+++ Magefile.go
@@ -5,7 +5,8 @@ package main
 
 import (
 	// mage:import
-	"github.com/grafana/grafana-plugin-sdk-go/build"
+        "github.com/grafana/grafana-plugin-sdk-go/build"
+        "github.com/magefile/mage/mg"
 )
 
 func init() {
@@ -16,5 +17,13 @@ func init() {
 	})
 }
 
+func Build() {
+	b := build.Build{}
+	linuxS390 := func () error {
+		return b.Custom("linux", "s390x")
+	}
+	mg.Deps(b.Linux, b.Windows, b.Darwin, b.DarwinARM64, b.LinuxARM64, b.LinuxARM, linuxS390)
+}
+
 // Default configures the default target.
-var Default = build.BuildAll
+var Default = Build